+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Pending
 
 - Change CLI name to `minhtml` as it's a more concise command name and allows for `cargo install minhtml`.
+- Add `keep_ssi_comments` to preserve SSI comments.
 
 ## 0.11.5